requirements.txt
numpy
luxai-s3
